Convert EncryptionEvent to Typescript
